+++ Common/Cutscenes/DialoguePlayer.cs
@@ -3,12 +3,11 @@ using Terraria.ModLoader;
 using System.Collections.Generic;
 using Microsoft.Xna.Framework;
 using Terraria;
+using System.Xml;
 
 namespace OvermorrowMod.Common.Cutscenes
 {
-    // Should be able to push dialogue that will constantly loop through each and pop them from the list
-    // Dialogue should be able to be cleared and then a new one added if it is urgent
-    public class Dialogue
+    public class Popup
     {
         public Texture2D speakerPortrait;
 
@@ -17,8 +16,8 @@ namespace OvermorrowMod.Common.Cutscenes
         public int showTime;
         public string bracketColor;
 
-        public bool openAnimation;
-        public bool closeAnimation;
+        private bool openAnimation;
+        private bool closeAnimation;
 
         /// <summary>
         /// Used to store information about a dialogue object
@@ -27,7 +26,7 @@ namespace OvermorrowMod.Common.Cutscenes
         /// <param name="drawTime">The amount of time it takes to completely draw the text</param>
         /// <param name="showTime">How long the text remains on screen after fully drawing</param>
         /// <param name="bracketColor">The hex color of the text when enclosed in brackets</param>
-        public Dialogue(Texture2D speakerPortrait, string displayText, int drawTime, int showTime, string bracketColor, bool openAnimation, bool closeAnimation)
+        public Popup(Texture2D speakerPortrait, string displayText, int drawTime, int showTime, string bracketColor, bool openAnimation, bool closeAnimation)
         {
             this.speakerPortrait = speakerPortrait;
             this.displayText = displayText;
@@ -38,29 +37,32 @@ namespace OvermorrowMod.Common.Cutscenes
             this.openAnimation = openAnimation;
             this.closeAnimation = closeAnimation;
         }
+
+        public bool ShouldOpen() => openAnimation;
+        public bool ShouldClose() => closeAnimation;
     }
 
     public class DialoguePlayer : ModPlayer
     {
-        private Queue<Dialogue> DialogueQueue = new Queue<Dialogue>();
-        public bool ShowDialogue = false;
+        private Queue<Popup> PopupQueue = new Queue<Popup>();
+
+        public bool AddedDialogue = false;
 
         public bool pickupWood = false;
         public bool distanceGuide = false;
         public bool guideGreeting = false;
 
-        public void AddDialogue(Texture2D speakerPortrait, string displayText, int drawTime, int showTime, Color bracketColor, bool openAnimation = true, bool closeAnimation = true)
+        public void AddPopup(Texture2D speakerPortrait, string displayText, int drawTime, int showTime, Color bracketColor, bool openAnimation = true, bool closeAnimation = true)
         {
-            ShowDialogue = true;
-            DialogueQueue.Enqueue(new Dialogue(speakerPortrait, displayText, drawTime, showTime, bracketColor.Hex3(), openAnimation, closeAnimation));
+            PopupQueue.Enqueue(new Popup(speakerPortrait, displayText, drawTime, showTime, bracketColor.Hex3(), openAnimation, closeAnimation));
         }
 
-        public Dialogue GetDialogue() => DialogueQueue.Peek();
+        public Popup GetPopup() => PopupQueue.Peek();
 
-        public void ClearDialogue() => DialogueQueue.Clear();
+        public void ClearPopup() => PopupQueue.Clear();
 
-        public void DequeueDialogue() => DialogueQueue.Dequeue();
+        public void RemovePopup() => PopupQueue.Dequeue();
 
-        public int GetQueueLength() => DialogueQueue.Count;
+        public int GetQueueLength() => PopupQueue.Count;
     }
 }
