+++ Common/Cutscenes/DialogueSystem.cs
@@ -8,16 +8,22 @@ namespace OvermorrowMod.Common.Cutscenes
 {
     public class DialogueSystem : ModSystem
     {
-        UserInterface DialogueUI;
-        internal DialogueState DialogueState;
+        UserInterface PopupUI;
+        internal PopupState PopupState;
 
+        //UserInterface DialogueUI;
+        //internal DialogueState DialogueState;
         public override void Load()
         {
             if (!Main.dedServ)
             {
-                DialogueState = new DialogueState();
+                PopupState = new PopupState();
+                PopupUI = new UserInterface();
+                PopupUI.SetState(PopupState);
+
+                /*DialogueState = new DialogueState();
                 DialogueUI = new UserInterface();
-                DialogueUI.SetState(DialogueState);
+                DialogueUI.SetState(DialogueState);*/
             }
         }
 
@@ -30,7 +36,8 @@ namespace OvermorrowMod.Common.Cutscenes
                     "OvermorrowMod: Dialogue",
                     delegate
                     {
-                        DialogueUI.Draw(Main.spriteBatch, new GameTime());
+                        PopupUI.Draw(Main.spriteBatch, new GameTime());
+                        //DialogueUI.Draw(Main.spriteBatch, new GameTime());
                         return true;
                     },
                     InterfaceScaleType.UI)
@@ -40,7 +47,8 @@ namespace OvermorrowMod.Common.Cutscenes
 
         public override void UpdateUI(GameTime gameTime)
         {
-            DialogueUI?.Update(gameTime);
+            PopupUI?.Update(gameTime);
+            //DialogueUI?.Update(gameTime);
         }
     }
 }
