+++ Common/OvermorrowGlobalNPC.cs
@@ -28,34 +28,6 @@ namespace OvermorrowMod.Common
 
         public int FungiTime;
 
-        public bool distanceCheck = false;
-
-        public override bool PreAI(NPC npc)
-        {
-            if (npc.type == NPCID.Guide)
-            {
-                foreach (Player player in Main.player)
-                {
-                    if (!player.active) continue;
-
-                    DialoguePlayer dialoguePlayer = player.GetModPlayer<DialoguePlayer>();
-
-                    if (dialoguePlayer.distanceGuide) return base.PreAI(npc);
-
-                    float xDistance = Math.Abs(npc.Center.X - player.Center.X);
-                    if (xDistance > 15 * 16)
-                    {
-                        dialoguePlayer.distanceGuide = true;
-
-                        Texture2D texture = ModContent.Request<Texture2D>(AssetDirectory.UI + "Portraits/Guide/GuideSmug", ReLogic.Content.AssetRequestMode.ImmediateLoad).Value;
-                        dialoguePlayer.AddDialogue(texture, "Where are you off to in such a hurry?", 60, 120, new Color(52, 201, 235), true, true);
-                    }
-                }
-            }
-
-            return base.PreAI(npc);
-        }
-
         public override void ResetEffects(NPC npc)
         {
             FungiInfection = false;
